Makes sender nullable

+++ lib/models/user.dart
@@ -34,8 +34,7 @@ class User extends AppEntity {
   factory User.fromMap(dynamic map) {
     if (map == null) throw ArgumentError('The type of user map is null');
 
-    // print('User: $map');
-
+    // Logger().d('User: $map');
     return User(
       uid: map['uid'],
       name: map['name'] ?? '',
@@ -44,8 +43,8 @@ class User extends AppEntity {
       role: map['role'] ?? 'default',
       status: map['status'] ?? 'offline',
       statusMessage: map['statusMessage'] ?? '',
-      lastActiveAt:
-          DateTime.fromMillisecondsSinceEpoch(map['lastActiveAt'] * 1000),
+      lastActiveAt: DateTime.fromMillisecondsSinceEpoch(
+          map['lastActiveAt'].toInt() * 1000),
       tags: List<String>.from(map['tags'] ?? []),
       hasBlockedMe: map['hasBlockedMe'] ?? false,
       blockedByMe: map['blockedByMe'] ?? false,
